refactor(e2e): rename e2e_p2p to p2p and delete reex
Now that the slashing/sentinel/governance tests have moved to multi-node/, e2e_p2p/
holds only real-libp2p transport tests, so rename the folder to p2p/ for category
consistency and delete reex.test.ts (pays setup cost for zero tests).

The src/e2e_!(prover)/*.test.ts auto-glob stops matching after the rename, so add
explicit src/p2p/*.test.ts globs (and repoint the reqresp glob) in both bootstrap.sh
test arrays, making the compat name derivation robust to the slash-bearing p2p path.
Repoint the blanket src/e2e_p2p/.* flake pattern to src/p2p/.*, the add_rollup harness
import, and the AUTOMINE preset doc comment. reqresp_no_handshake stays until its flake
is fixed.
